Move at::internal_set_names_inplace to ATen/NamedTensorUtils.h
I'm planning on dumping helper functions here in the near future.

Test Plan:
- [namedtensor ci]
